pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sam_loss"  # DON'T FORGET TO REMOVE empty FROM git remote!!!
version = "0.0.1"
description = "<<<ENTER_DESCRIPTION>>>"
authors = [
    { name = "Tobias Kirschstein", email = "tobias.kirschstein@gmail.com" },
]
readme = "README.md"
license = { text = "Apache 2.0" }
requires-python = ">=3.8.0"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
# urls = { Documentation = "<<<ENTER_LINK_TO_DOCUMENTATION>>>" }
# Main dependencies
dependencies = [
]

[project.optional-dependencies]
# Development packages, install via <<<PROJECT_NAME>>>[dev]
dev = [
    'hydra-core',
    'omegaconf',
    'iopath'
]

[project.scripts]
# E.g., ns-download-data = "scripts.downloads.download_data:entrypoint"

[tool.setuptools.packages.find]
where = ["src"]
include = ["sam_loss*"]  # Keep the '*', otherwise submodules are not found
